+++ lib/core/dirinit.dart
@@ -25,7 +25,7 @@ Future<void> initDirectoryData() async {
     await Directory(realDataPath).create();
   }
   for (final String file in files) {
-    // TODO: should retain data in the future for direct past paper list updates
+    // TODO: [Micfong] should retain data in the future for direct past paper list updates
     // if (!await File('$realDataPath$file').exists()) {
     copied = true;
     final String data = await rootBundle.loadString('$bundledDataPath$file');