+++ .github/docs/release-notes.md
@@ -4,6 +4,7 @@
   - Filtering by subjects and / or by levels
   - Checking answers with one click
   - Previewing original question papers and mark schemes
+- Enhances application performance
 
 **Note:** Past Paper Master is still in beta test, and there are still many features to be added and bugs to be fixed. Please report any issues you encounter to SCIE.DEV. Thank you for your support!
 
@@ -13,9 +14,9 @@ See [commit history](https://github.com/SCIEDEV/PastPaperMaster/commits/main) fo
 
 Currently, none of the binaries are signed due to a lack of developer certificate. (which is expensive!)
 
-- `ppm-macos-v0.1.7-beta+72.tar.gz` macOS Universal application.
+- `ppm-macos-v0.2.0-beta+76.tar.gz` macOS Universal application.
   - Since this application is not signed, you may need to click the <kbd>Open anyway</kbd> button in System Settings » Privacy & Security.
   - Opening this app anyway should be secure since this app is already open source; if you are still concerned of security, just build this app from source and use that binary.
-- `ppm-windows-v0.1.7-beta+72.zip` Windows x64 application.
+- `ppm-windows-v0.2.0-beta+76.zip` Windows x64 application.
   - Again, this application is not signed. You may see Windows Defender warning about this application.
-- `ppm-linux-v0.1.7-beta+72.tar.gz` Linux x64 application.
+- `ppm-linux-v0.2.0-beta+76.tar.gz` Linux x64 application.